+++ src/Contracts/NotificationInterface.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace Ntavelis\Mercure\Contracts;
 
-interface NotificationInterface extends \JsonSerializable
+interface NotificationInterface
 {
     public function getTokenData(): array;
+
+    public function toArray(): array;
 }