+++ src/Messages/PrivateNotification.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Ntavelis\Mercure\Messages;
 
+use Ntavelis\Mercure\Config\ConfigStamp;
 use Ntavelis\Mercure\Contracts\NotificationInterface;
 
 class PrivateNotification implements NotificationInterface
@@ -11,12 +12,14 @@ class PrivateNotification implements NotificationInterface
     private $topics;
     private $data;
     private $targets;
+    private $configStamp;
 
     public function __construct(array $topics, array $data, array $targets)
     {
         $this->topics = $topics;
         $this->data = $data;
         $this->targets = $targets;
+        $this->configStamp = new ConfigStamp();
     }
 
     public function getTopics(): array
@@ -34,20 +37,30 @@ class PrivateNotification implements NotificationInterface
         return $this->targets;
     }
 
+    public function getConfigStamp(): ConfigStamp
+    {
+        return $this->configStamp;
+    }
+
+    public function withConfig(ConfigStamp $configStamp): void
+    {
+        $this->configStamp = $configStamp;
+    }
+
     public function getTokenData(): array
     {
         return array_merge($this->getTopics(), $this->getTargets());
     }
 
-    /**
-     * @inheritDoc
-     */
-    public function jsonSerialize(): array
+    public function toArray(): array
     {
         return [
             'topic' => $this->getTopics(),
             'data' => json_encode($this->getData()),
             'target' => $this->getTargets(),
+            'id' => $this->configStamp->getId(),
+            'type' => $this->configStamp->getType(),
+            'retry' => $this->configStamp->getRetry(),
         ];
     }
 }