+++ tests/Unit/Messages/PrivateNotificationTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Ntavelis\Mercure\Tests\Unit\Messages;
 
+use Ntavelis\Mercure\Config\ConfigStamp;
 use Ntavelis\Mercure\Messages\PrivateNotification;
 use PHPUnit\Framework\TestCase;
 
@@ -28,6 +29,29 @@ class PrivateNotificationTest extends TestCase
             'topic' => ['topics'],
             'data' => '["data"]',
             'target' => ['ntavelis'],
-        ], $privateNotification->jsonSerialize());
+            'id' => null,
+            'type' => null,
+            'retry' => null,
+        ], $privateNotification->toArray());
+    }
+
+    /** @test */
+    public function itCanBeConfiguredThroughAConfigStampClass(): void
+    {
+        $notification = new PrivateNotification(['topics'], ['data'], ['ntavelis']);
+        $configStamp = (new ConfigStamp())
+            ->setType('notification')
+            ->setId('12314')
+            ->setRetry(30);
+        $notification->withConfig($configStamp);
+
+        $this->assertSame([
+            'topic' => ['topics',],
+            'data' => '["data"]',
+            'target' => ['ntavelis'],
+            'id' => '12314',
+            'type' => 'notification',
+            'retry' => 30,
+        ], $notification->toArray());
     }
 }