+++ CHANGELOG.md
@@ -4,7 +4,18 @@ All notable changes to `ntavelis/mercure-php` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## [Unreleased]
+## [Unreleased
+
+[0.4.0] - 2020-05-16
+
+### Added
+- Added a ConfigStamp class, that accepts specific config values for the notifications
+- Added the ability to pass the config stamp through the fluent builders
+
+### Changed
+- Changed the NotificationInterface no longer extends jsonSerializable, instead it exposes a toArray method
+- Interface docblock, now it contains proper exception in docblock 
+- Updated the README.md
 
 ## [0.3.2] - 2020-05-09